Add files via upload
Modeled solution for a Bernoulli multi-arm bandit problem. 
An advertising firm has 20 customers, and there is an unknown probability p_i (p_1, p_2, ...., p_20) for each customer that their advertisement is clicked. The code simulates the 20 probabilities using the beta distribution, and then implements different reinforcement learning strategies to maximize revenue. The reinforcement learning algorithms used include epsilon-greedy, adaptive-epsilon greedy (pure exploration for the first X runs, then pure exploitation for the rest of the runs) as well as a strategy that decreases epsilon with time. The script also implements strategies using optimistic initial values, UCB, and the gradient method (with the softmax function). For each strategy, the rewards are plotted along with the regret per round for each strategy.