Add files via upload
Markov Decision Processes and Dynamic Programming
The script models a ski-shop inventory problem, in which a sporting goods store can buy and stock skis from the beginning of October to the beginning of March. The shop can store up to 100 skis, each set costs $500 and is sold for $700, and the owner must decide how many skis to buy at the beginning of each month (the total number of skis in the shop can never exceed 100). Any skis remaining at the end of March are sold for $300. 
Given that demand for skis follows a Poisson distribution with the following average demand per month, the script aims to find the optimal inventory buying policy using Markov Decision Processes and dynamic programming. The optimal policy (and profits attained) is then graphically compared to a random policy. 

Average demand:
October - 20
November - 40
December - 50
January - 35
February - 25
March - 5